wind-monitor-app/backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.11.2
python-dateutil==2.9.0
